サイドバーのブックマーク一覧での「すべてのブックマークを削除」を実装。

+++ chrome/content/common/31-Model-Bookmark.js
@@ -101,9 +101,9 @@ extend(Bookmark, {
         return res;
     },
     deleteBookmarks: function(bookmarks) {
-        bookmarks = [].concat(bookmarks);
         let bmIds = bookmarks.filter(function (b) b.id > 0)
                              .map(function (b) b.id);
+        if (!bmIds.length) return;
         let placeholder = bmIds.map(function () "?").join(",");
         Model.Tag.db.execute("delete from tags where bookmark_id in (" +
                              placeholder + ")", bmIds);
